% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecometrics.R
\name{ACF_PLOT}
\alias{ACF_PLOT}
\title{Plots ACF of a univariate time series}
\usage{
ACF_PLOT(x, lag.max = NULL)
}
\arguments{
\item{x}{numeric vector}

\item{lag.max}{maximum lag to calculate the acf}
}
\value{
a plot of the acf vs lag
}
\description{
Plots ACF of a univariate time series
}
\examples{
data(keconomy)
attach(keconomy)
ACF_PLOT(UR)



}
\author{
Mutua Kilai
}
