% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecometrics.R
\name{autocorrelation_assumption}
\alias{autocorrelation_assumption}
\title{Check model for residual independence}
\usage{
autocorrelation_assumption(model)
}
\arguments{
\item{model}{A lm object}
}
\value{
returns the p-value for the test
}
\description{
Checks model for independence of residuals
}
\examples{
model <- lm(pi ~ hs + ps, data = eduperform)
autocorrelation_assumption(model)
}
\author{
Mutua Kilai
}
