% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecometrics.R
\name{best_arima}
\alias{best_arima}
\title{Select Optimal Model based on BIC}
\usage{
best_arima(data, max_p = 5, max_d = 2, max_q = 5)
}
\arguments{
\item{data}{A univariate ts object}

\item{max_p}{Maximum AR order}

\item{max_d}{Maximum differencing order}

\item{max_q}{Maximum MA order}
}
\value{
A list containing the optimal model results and the BIC value
}
\description{
Select Optimal Model based on BIC
}
\examples{
data(keconomy)
attach(keconomy)
best_arima(UR, max_p = 5, max_d = 2, max_q = 5)

}
