% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecometrics.R
\name{check_model_sig}
\alias{check_model_sig}
\title{Checking Overall Model Significance}
\usage{
check_model_sig(data, y, x)
}
\arguments{
\item{data}{A data frame containing the variables to use}

\item{y}{The dependent variable}

\item{x}{A set of independent variables}
}
\value{
p-value with a statement on whether the model is significant or not
}
\description{
Checking Overall Model Significance
}
\examples{
check_model_sig(data = eduperform, "pi", c("hs", "ps"))


}
\author{
Mutua Kilai
}
