% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecometrics.R
\name{check_stationarity}
\alias{check_stationarity}
\title{Check Series for Weak Stationarity}
\usage{
check_stationarity(x)
}
\arguments{
\item{x}{A numeric vector or time series object}
}
\value{
p-value of the test
}
\description{
Check Series for Weak Stationarity
}
\examples{
data(keconomy)
attach(keconomy)
check_stationarity(UR)

}
\author{
Mutua Kilai
}
