% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eduperform.R
\docType{data}
\name{eduperform}
\alias{eduperform}
\title{Student Performance Data}
\format{
\subsection{\code{eduperform}}{

A data frame with 10000 rows and 6 columns:
\describe{
\item{hs}{hours studied}
\item{ps}{previous score}
\item{ea}{extracurricula activities}
\item{sh}{sleep hours}
\item{sqpp}{sample question paper practiced}
\item{pi}{performance Index}
...
}
}
}
\source{
\url{https://www.kaggle.com/datasets/nikhil7280/student-performance-multiple-linear-regression?resource=download}
}
\usage{
eduperform
}
\description{
Student performance dataset is a dataset designed to examine the factors influencing academic student performance.
}
\keyword{datasets}
