% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecometrics.R
\name{fit_arima}
\alias{fit_arima}
\title{Fit ARIMA models to univariate data}
\usage{
fit_arima(data, p, d, q)
}
\arguments{
\item{data}{a univariate class object or a vector}

\item{p}{AR order}

\item{d}{differencing order}

\item{q}{MA order}
}
\value{
A tibble containing the estimate, SE and p-value
}
\description{
Fit ARIMA models to univariate data
}
\examples{
data(keconomy)
attach(keconomy)
fit_arima(UR, p=2,d=0,q=3)

}
