% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecometrics.R
\name{get_coefficients_variance}
\alias{get_coefficients_variance}
\title{Get variance of the model coefficients}
\usage{
get_coefficients_variance(data, y, x)
}
\arguments{
\item{data}{A data frame containing the variables to use}

\item{y}{The dependent variable}

\item{x}{A set of independent variables}
}
\value{
Tibble containing the variances
}
\description{
Get variance of the model coefficients
}
\examples{
get_coefficients_variance(data = eduperform, "pi", c("hs", "ps"))


}
\author{
Mutua Kilai
}
