% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecometrics.R
\name{get_confint}
\alias{get_confint}
\title{Confidence Intervals of Model Parameters}
\usage{
get_confint(data, y, x, level = 0.95)
}
\arguments{
\item{data}{A data frame containing the variables to use}

\item{y}{The dependent variable}

\item{x}{A set of independent variables}

\item{level}{level of significance can be 0.95, 0.90 etc. default is 0.95}
}
\value{
tibble containing the lower and upper confidence intervals
}
\description{
Confidence Intervals of Model Parameters
}
\examples{
get_confint(data = eduperform, "pi", c("hs", "ps"))


}
\author{
Mutua Kilai
}
