% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecometrics.R
\name{get_significant_predictors}
\alias{get_significant_predictors}
\title{Obtaining only significant predictors from a model}
\usage{
get_significant_predictors(data, y, x, alpha = 0.05)
}
\arguments{
\item{data}{A data frame containing the variables to use}

\item{y}{The dependent variable}

\item{x}{A set of independent variables}

\item{alpha}{desired alpha level. default is 0.05}
}
\value{
A tibble containing the significant predictors
}
\description{
Obtaining only significant predictors from a model
}
\examples{
get_significant_predictors(data = eduperform, "pi", c("hs", "ps"))



}
\author{
Mutua Kilai
}
