% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecometrics.R
\name{heteroscedasticity_assumption}
\alias{heteroscedasticity_assumption}
\title{Checking heteroscedasticity assumption}
\usage{
heteroscedasticity_assumption(model)
}
\arguments{
\item{model}{A lm model object}
}
\value{
The p-value of the test statistic.
}
\description{
Checking heteroscedasticity assumption
}
\examples{
model <- lm(pi ~ hs + ps, data = eduperform)
heteroscedasticity_assumption(model)

}
\author{
Mutua Kilai
}
