% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keconomy.R
\docType{data}
\name{keconomy}
\alias{keconomy}
\title{Kenya Unemployment Rate and GDP Growth rate for 1999-2023}
\format{
\subsection{\code{keconomy}}{

A data frame with 25 rows and 3 columns:
\describe{
\item{Year}{Year; from 1999 to 2023}
\item{UR}{Unemployment Rate}
\item{GR}{GDP Growth Rate}
}
}
}
\source{
\url{https://www.statista.com}
}
\usage{
keconomy
}
\description{
Annual Time Series data for Kenyan Economy showing the unemployment rate and GDP Growth Rate.
}
\keyword{datasets}
