% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecometrics.R
\name{multicollinearity_assumption}
\alias{multicollinearity_assumption}
\title{Multicollinearity Assumption}
\usage{
multicollinearity_assumption(model)
}
\arguments{
\item{model}{A lm object}
}
\value{
A tibble containing the VIF and Tolerance values
}
\description{
Multicollinearity Assumption
}
\examples{
model <- lm(pi ~ hs + ps, data = eduperform)
multicollinearity_assumption(model)


}
\author{
Mutua Kilai
}
