% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecometrics.R
\name{normality_assumption}
\alias{normality_assumption}
\title{Checking normality of residuals}
\usage{
normality_assumption(model)
}
\arguments{
\item{model}{A lm model object}
}
\value{
The p-value of the test statistic.
}
\description{
Checking normality of residuals
}
\examples{
model <- lm(pi ~ hs + ps, data = eduperform)
normality_assumption(model)

}
\author{
Mutua Kilai
}
