% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecometrics.R
\name{ols_model}
\alias{ols_model}
\title{Fitting a simple or multiple linear regression}
\usage{
ols_model(data, y, x)
}
\arguments{
\item{data}{A data frame containing the variables to use}

\item{y}{The dependent variable}

\item{x}{Set of independent variables}
}
\value{
A tibble of the coefficients, standard errors, t-statistics and p-value
}
\description{
Fitting a simple or multiple linear regression
}
\examples{
ols_model(data = eduperform, "pi", c("hs", "ps"))


}
\author{
Mutua Kilai
}
