% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecometrics.R
\name{ols_model_sig}
\alias{ols_model_sig}
\title{F-statistic attributes}
\usage{
ols_model_sig(data, y, x)
}
\arguments{
\item{data}{A data frame containing the variables to use}

\item{y}{The dependent variable}

\item{x}{Set of independent variables}
}
\value{
A tibble containing the number of observations, F-Statistic, degrees of freedom and p-value
}
\description{
F-statistic attributes
}
\examples{
ols_model_sig(data = eduperform, "pi", c("hs", "ps"))



}
\author{
Mutua Kilai
}
