% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecometrics.R
\name{ols_model_stats}
\alias{ols_model_stats}
\title{Model Summary Statistics}
\usage{
ols_model_stats(data, y, x)
}
\arguments{
\item{data}{A data frame containing the variables to use}

\item{y}{The dependent variable}

\item{x}{The independent variables}
}
\value{
A tibble containing model summary stats: R-Squared, Adjusted R-Squared, AIC and BIC
}
\description{
Model Summary Statistics
}
\examples{
ols_model_stats(data = eduperform, "pi", c("hs", "ps"))



}
\author{
Mutua Kilai
}
