% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecometrics.R
\name{PACF_PLOT}
\alias{PACF_PLOT}
\title{Plots PACF of a univariate time series}
\usage{
PACF_PLOT(x, lag.max = NULL)
}
\arguments{
\item{x}{a numeric vector}

\item{lag.max}{maximum lag to calculate pacf}
}
\value{
a plot of the pacf vs lag
}
\description{
Plots PACF of a univariate time series
}
\examples{
data(keconomy)
attach(keconomy)
PACF_PLOT(UR)



}
\author{
Mutua Kilai
}
